+++ src/client/java/tempeststudios/inventorysort/core/InventorySortCore.java
@@ -1,0 +1,12 @@
+package tempeststudios.inventorysort.core;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public final class InventorySortCore {
+    public static final String MOD_ID = "inventorysort";
+    public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
+
+    private InventorySortCore() {
+    }
+}